drivers: usb: gadget: Fix usb tethering by enforcing ncm
For a thus far unknown reason, we hit an npe when using rndis without multi packet transfer.

More specifically, the npe is cause by the following memcpy:

/* copy skb data */
memcpy(req->buf, skb->data,
	skb->len);

I will revisit it later to see if it can be fixed properly. But tests show that sharing WiFi and mobile data now works both on Windows and Linux, so let's fix it for now.